# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/property-types": {
      "get": {
        "operationId": "PropertyTypesController_list_api/v1",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true (por defecto): sólo activos, los de los selectores. false: sólo desactivados. all: todos.",
            "schema": {
              "default": "true",
              "type": "string",
              "enum": [
                "true",
                "false",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogItemDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar (por defecto, sólo los activos)",
        "tags": [
          "Master data"
        ]
      },
      "post": {
        "operationId": "PropertyTypesController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Nombre repetido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/property-types/{id}": {
      "get": {
        "operationId": "PropertyTypesController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver uno",
        "tags": [
          "Master data"
        ]
      },
      "patch": {
        "operationId": "PropertyTypesController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Nombre repetido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar el nombre (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/property-types/{id}/deactivate": {
      "patch": {
        "description": "Sale de los selectores; lo que ya lo usa lo sigue mostrando. No se borra nunca.",
        "operationId": "PropertyTypesController_deactivate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/property-types/{id}/activate": {
      "patch": {
        "operationId": "PropertyTypesController_activate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/amenities": {
      "post": {
        "operationId": "AmenitiesController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAmenityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Nombre repetido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear (sólo admin)",
        "tags": [
          "Master data"
        ]
      },
      "get": {
        "operationId": "AmenitiesController_list_api/v1",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true (por defecto): sólo activos, los de los selectores. false: sólo desactivados. all: todos.",
            "schema": {
              "default": "true",
              "type": "string",
              "enum": [
                "true",
                "false",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogItemDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar (por defecto, sólo los activos)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/amenities/{id}": {
      "patch": {
        "operationId": "AmenitiesController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAmenityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Nombre repetido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar nombre o ícono (sólo admin)",
        "tags": [
          "Master data"
        ]
      },
      "get": {
        "operationId": "AmenitiesController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver uno",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/amenities/{id}/deactivate": {
      "patch": {
        "description": "Sale de los selectores; lo que ya lo usa lo sigue mostrando. No se borra nunca.",
        "operationId": "AmenitiesController_deactivate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/amenities/{id}/activate": {
      "patch": {
        "operationId": "AmenitiesController_activate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/operation-types": {
      "get": {
        "operationId": "OperationTypesController_list_api/v1",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true (por defecto): sólo activos, los de los selectores. false: sólo desactivados. all: todos.",
            "schema": {
              "default": "true",
              "type": "string",
              "enum": [
                "true",
                "false",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogItemDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar (por defecto, sólo los activos)",
        "tags": [
          "Master data"
        ]
      },
      "post": {
        "operationId": "OperationTypesController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Nombre repetido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/operation-types/{id}": {
      "get": {
        "operationId": "OperationTypesController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver uno",
        "tags": [
          "Master data"
        ]
      },
      "patch": {
        "operationId": "OperationTypesController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Nombre repetido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar el nombre (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/operation-types/{id}/deactivate": {
      "patch": {
        "description": "Sale de los selectores; lo que ya lo usa lo sigue mostrando. No se borra nunca.",
        "operationId": "OperationTypesController_deactivate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/operation-types/{id}/activate": {
      "patch": {
        "operationId": "OperationTypesController_activate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
...
    "/api/v1/service-types/{id}/deactivate": {
      "patch": {
        "description": "Sale de los selectores; lo que ya lo usa lo sigue mostrando. No se borra nunca.",
        "operationId": "ServiceTypesController_deactivate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
    "/api/v1/service-types/{id}/activate": {
      "patch": {
        "operationId": "ServiceTypesController_activate_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDto"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivar (sólo admin)",
        "tags": [
          "Master data"
        ]
      }
    },
...
  },
  "info": {
    "title": "Adminprop API",
    "description": "API NestJS con arquitectura hexagonal, CQRS y autenticación JWT propia.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token devuelto por POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "CatalogItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3f8c1b1e-3a6f-4f2e-9a1c-6d9f2b7c4e11"
          },
          "name": {
            "type": "string",
            "example": "Departamento"
          },
          "icon": {
            "type": "string",
            "nullable": true,
            "description": "Sólo amenities: nombre de ícono de lucide. En el resto, null.",
            "example": null
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-19T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-19T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "icon",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCatalogItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Dúplex",
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCatalogItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Dúplex",
            "maxLength": 120
          }
        }
      },
...
    }
  }
}